requirements.txt
python-telegram-bot==13.7
pandas==2.2.2
openpyxl==3.1.2
requests==2.31.0
cryptography==42.0.5
flask==3.0.3